requirements_dev.txt
# project requirements below
pandas
dill
scikit-learn
pymongo

# dev requirements -
pytest==7.1.3
tox==3.25.1
black==22.8.0
flake8==5.0.4
mypy==0.971
notebook
matplotlib
mpl-tools
seaborn
python-dotenv
-e .